refactor: move ActiveRecord support under its own namespace
ActiveRecordSupport becomes TranslationDiff::ActiveRecord::Support,
the mixin the cache store and rate limiter both include. Since
TranslationDiff::ActiveRecord now shares its name with ::ActiveRecord,
the two references it makes to the real gem across a genuinely
missing constant (ar_error?'s defined?/is_a? pair) are :: qualified;
the version floor check and the anonymous model's base class already
were.
