+++ lib/translation_diff/active_record.rb
@@ -1,0 +1,3 @@
+# The namespace the cache store and the rate limiter's ActiveRecord support live under; see active_record/support.
+module TranslationDiff::ActiveRecord
+end
